fund-flow-tracker/backend/requirements.txt
fastapi
uvicorn
neo4j
supabase
faker
numpy
presidio-analyzer
presidio-anonymizer
google-generativeai
langchain
langchain-google-genai
reportlab
pillow
python-jose
passlib
python-multipart
pydantic
python-dotenv
